Added in a Resource Handler
This resource handler uses the Singleton pattern to maintain a single instance of our resource handler that is globally available accross our application.

The handler itself contains a resource holder for each of its supported resource types (textures, fonts and sounds) along with a mutex to mutually exclude access to each resource holder to maintain thread safety, this however does not garantee thread safety to the resource itself.

Note: sf::Font is modified whenever it is attached to a sfml text object and therefore is susceptible to data race.